requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==3.2.24
django-jazzmin==2.6.1
django-tinymce==3.6.1
django_debug_toolbar==3.8.1
gunicorn==22.0.0
idna==3.7
importlib-metadata==6.7.0
packaging==24.0
Pillow==9.5.0
python-decouple==3.8
python-dotenv==0.21.1
pytz==2023.3.post1
requests==2.31.0
sqlparse==0.4.4
typing_extensions==4.7.1
urllib3==2.0.7
zipp==3.15.0
